Force global namespace for utility types and ignore incorrect generic use
To reduce the likelyhood of conflicts with source code we only support
using `\` global namespaces in utility types.

If the incorrect number of generics is passed, return `null` instead of
error in case of global classes existing.

Not perfect as there is still the possibility of global classes which
have the same number of generics but greatly reduces as only the global
namespace could conflict instead of every namespace.